% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_SQL_db.R
\name{build_SQL_db}
\alias{build_SQL_db}
\title{Build SQL database}
\usage{
build_SQL_db(db_directory = ".", database = NULL)
}
\arguments{
\item{db_directory}{directory where files are held}

\item{database}{database file from RSQLite}
}
\value{
nothing
}
\description{
Build SQL database
}
\examples{
build_SQL_db()

}
