% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_data.R
\name{get_census_data}
\alias{get_census_data}
\title{Create census dataset}
\usage{
get_census_data(
  db_directory = ".",
  census_tract_files = NULL,
  census_zipcode_files = NULL
)
}
\arguments{
\item{db_directory}{path to directory where data files are held}

\item{census_tract_files}{vector of names of census tract files}

\item{census_zipcode_files}{vector of names of census zipcode files}
}
\value{
a list of dataframes from each given census file
}
\description{
Function will identify all of the census datasets in your directory and create a dataset for them
}
\examples{
hospitalDB <- SEERdb::get_hospital_data("../../master data files - SEER Medicare 2024/")


hospitalDB <- SEERdb::get_hospital_data(
                "../../master data files - SEER Medicare 2024/",
                c("hospital1996.withzip.txt.gz", "hospital2000.withzip.txt.gz")
                )

}
