% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dme_data.R
\name{get_dme_data}
\alias{get_dme_data}
\title{Create durable medical equipment dataset}
\usage{
get_dme_data(
  db_directory = ".",
  dme_base_files = NULL,
  dme_demo_files = NULL,
  dme_line_files = NULL
)
}
\arguments{
\item{db_directory}{path to directory where data files are held}

\item{dme_base_files}{vector of names of hsp.base files}

\item{dme_demo_files}{vector of names of hsp.condition files}

\item{dme_line_files}{vector of names of hsp.occurence files}
}
\value{
a list of dataframes from each given hsp file
}
\description{
Function will identify all of the hospice claims datasets in your directory and create a dataset for them.
There are 6 segments provided: base, revenue, condition, occurence, span, and value
}
\details{
PATIENT_ID and CLM_ID should be used for merging files
}
\examples{
hospitalDB <- SEERdb::get_hospital_data("../../master data files - SEER Medicare 2024/")


hospitalDB <- SEERdb::get_hospital_data(
                "../../master data files - SEER Medicare 2024/",
                c("hospital1996.withzip.txt.gz", "hospital2000.withzip.txt.gz")
                )

}
