% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hospital_data.R
\name{get_hospital_data}
\alias{get_hospital_data}
\title{Create hospital dataset}
\usage{
get_hospital_data(db_directory = ".", hospital_files = NULL)
}
\arguments{
\item{db_directory}{path to directory where data files are held}

\item{hospital_files}{vector of names of hospital files}
}
\description{
Function will identify all of the hospital datasets in your directory and create a dataset for them
}
