% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mbsf_data.R
\name{get_mbsf_data}
\alias{get_mbsf_data}
\title{Create mbsf dataset}
\usage{
get_mbsf_data(
  db_directory = ".",
  ab.summary_files = NULL,
  abcd.summary_files = NULL,
  cc.summary_files = NULL,
  oth.cc.summary_files = NULL
)
}
\arguments{
\item{db_directory}{path to directory where data files are held}

\item{ab.summary_files}{vector of names of ab summary files}

\item{abcd.summary_files}{vector of names of abcd summary files}

\item{cc.summary_files}{vector of names of cc summary files}

\item{oth.cc.summary_files}{vector of names of other cc summary files}
}
\value{
a list of dataframes from each given mbsf file
}
\description{
Function will identify all of the mbsf datasets in your directory and create a dataset for them
}
\examples{
hospitalDB <- SEERdb::get_hospital_data("../../master data files - SEER Medicare 2024/")


hospitalDB <- SEERdb::get_hospital_data(
                "../../master data files - SEER Medicare 2024/",
                c("hospital1996.withzip.txt.gz", "hospital2000.withzip.txt.gz")
                )

}
