% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nch_data.R
\name{get_nch_data}
\alias{get_nch_data}
\title{Create nch dataset}
\usage{
get_nch_data(
  db_directory = ".",
  nch_base_files = NULL,
  nch_demo_files = NULL,
  nch_line_files = NULL
)
}
\arguments{
\item{db_directory}{path to directory where data files are held}

\item{nch_base_files}{vector of names of nch base files}

\item{nch_demo_files}{vector of names of nch demo files}

\item{nch_line_files}{vector of names of nch line files}
}
\value{
a list of dataframes from each given hospital file
}
\description{
Function will identify all of the nch datasets in your directory and create a dataset for them
}
\examples{
hospitalDB <- SEERdb::get_hospital_data("../../master data files - SEER Medicare 2024/")


hospitalDB <- SEERdb::get_hospital_data(
                "../../master data files - SEER Medicare 2024/",
                c("hospital1996.withzip.txt.gz", "hospital2000.withzip.txt.gz")
                )

}
