% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SEER_data.R
\name{get_SEER_data}
\alias{get_SEER_data}
\title{Create SEER dataset}
\usage{
get_SEER_data(db_directory = ".", SEER_files = NULL)
}
\arguments{
\item{db_directory}{path to directory where data files are held}

\item{SEER_files}{vector of names of SEER files}
}
\value{
a list of dataframes from each given SEER file
}
\description{
Function will identify all of the SEER datasets in your directory and create a dataset for them
}
\examples{
hospitalDB <- SEERdb::get_hospital_data("../../master data files - SEER Medicare 2024/")


hospitalDB <- SEERdb::get_hospital_data(
                "../../master data files - SEER Medicare 2024/",
                c("hospital1996.withzip.txt.gz", "hospital2000.withzip.txt.gz")
                )

}
