% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospital_file_labels}
\alias{hospital_file_labels}
\title{Hospital File Labels.}
\format{
A data frame with four variables: \code{Start}, \code{Stop},
\code{name}, \code{label}
}
\usage{
hospital_file_labels
}
\description{
Guide for hospital files that include positions for values, column names and
column labels
}
\keyword{datasets}
